+++ src/Coroutine/Coroutine.php
@@ -3,7 +3,7 @@
 namespace Recoil\Coroutine;
 
 use Exception;
-use Recoil\Kernel\Strand\StrandInterface;
+use Recoil\Kernel\Strand\Strand;
 
 /**
  * A coroutine represents a unit of work that can be suspended and resumed.
@@ -13,39 +13,39 @@ interface Coroutine
     /**
      * Start the coroutine.
      *
-     * @param StrandInterface $strand The strand that is executing the coroutine.
+     * @param Strand $strand The strand that is executing the coroutine.
      */
-    public function call(StrandInterface $strand);
+    public function call(Strand $strand);
 
     /**
      * Resume execution of a suspended coroutine by passing it a value.
      *
-     * @param StrandInterface $strand The strand that is executing the coroutine.
-     * @param mixed           $value  The value to send to the coroutine.
+     * @param Strand $strand The strand that is executing the coroutine.
+     * @param mixed  $value  The value to send to the coroutine.
      */
-    public function resumeWithValue(StrandInterface $strand, $value);
+    public function resumeWithValue(Strand $strand, $value);
 
     /**
      * Resume execution of a suspended coroutine by passing it an exception.
      *
-     * @param StrandInterface $strand    The strand that is executing the coroutine.
-     * @param Exception       $exception The exception to send to the coroutine.
+     * @param Strand    $strand    The strand that is executing the coroutine.
+     * @param Exception $exception The exception to send to the coroutine.
      */
-    public function resumeWithException(StrandInterface $strand, Exception $exception);
+    public function resumeWithException(Strand $strand, Exception $exception);
 
     /**
      * Inform the coroutine that the executing strand is being terminated.
      *
-     * @param StrandInterface $strand The strand that is executing the coroutine.
+     * @param Strand $strand The strand that is executing the coroutine.
      */
-    public function terminate(StrandInterface $strand);
+    public function terminate(Strand $strand);
 
     /**
      * Finalize the coroutine.
      *
      * This method is invoked after the coroutine is popped from the call stack.
      *
-     * @param StrandInterface $strand The strand that is executing the coroutine.
+     * @param Strand $strand The strand that is executing the coroutine.
      */
-    public function finalize(StrandInterface $strand);
+    public function finalize(Strand $strand);
 }