+++ src/Coroutine/CoroutineAdaptor.php
@@ -3,7 +3,7 @@
 namespace Recoil\Coroutine;
 
 use InvalidArgumentException;
-use Recoil\Kernel\Strand\StrandInterface;
+use Recoil\Kernel\Strand\Strand;
 
 /**
  * Adapts arbitrary values into coroutine objects.
@@ -13,11 +13,11 @@ interface CoroutineAdaptor
     /**
      * Adapt a value into a coroutine.
      *
-     * @param StrandInterface $strand The currently executing strand.
-     * @param mixed           $value  The value to adapt.
+     * @param Strand $strand The currently executing strand.
+     * @param mixed  $value  The value to adapt.
      *
      * @return Coroutine
      * @throws InvalidArgumentException if now valid adaptation can be made.
      */
-    public function adapt(StrandInterface $strand, $value);
+    public function adapt(Strand $strand, $value);
 }