+++ src/Coroutine/CoroutineTrait.php
@@ -3,7 +3,7 @@
 namespace Recoil\Coroutine;
 
 use Exception;
-use Recoil\Kernel\Strand\StrandInterface;
+use Recoil\Kernel\Strand\Strand;
 
 /**
  * A trait for coroutines that provides basic default implementations.
@@ -13,17 +13,17 @@ trait CoroutineTrait
     /**
      * Start the coroutine.
      *
-     * @param StrandInterface $strand The strand that is executing the coroutine.
+     * @param Strand $strand The strand that is executing the coroutine.
      */
-    abstract public function call(StrandInterface $strand);
+    abstract public function call(Strand $strand);
 
     /**
      * Resume execution of a suspended coroutine by passing it a value.
      *
-     * @param StrandInterface $strand The strand that is executing the coroutine.
-     * @param mixed           $value  The value to send to the coroutine.
+     * @param Strand $strand The strand that is executing the coroutine.
+     * @param mixed  $value  The value to send to the coroutine.
      */
-    public function resumeWithValue(StrandInterface $strand, $value)
+    public function resumeWithValue(Strand $strand, $value)
     {
         $strand->returnValue($value);
     }
@@ -31,10 +31,10 @@ trait CoroutineTrait
     /**
      * Resume execution of a suspended coroutine by passing it an exception.
      *
-     * @param StrandInterface $strand    The strand that is executing the coroutine.
-     * @param Exception       $exception The exception to send to the coroutine.
+     * @param Strand    $strand    The strand that is executing the coroutine.
+     * @param Exception $exception The exception to send to the coroutine.
      */
-    public function resumeWithException(StrandInterface $strand, Exception $exception)
+    public function resumeWithException(Strand $strand, Exception $exception)
     {
         $strand->throwException($exception);
     }
@@ -42,9 +42,9 @@ trait CoroutineTrait
     /**
      * Inform the coroutine that the executing strand is being terminated.
      *
-     * @param StrandInterface $strand The strand that is executing the coroutine.
+     * @param Strand $strand The strand that is executing the coroutine.
      */
-    public function terminate(StrandInterface $strand)
+    public function terminate(Strand $strand)
     {
     }
 
@@ -53,9 +53,9 @@ trait CoroutineTrait
      *
      * This method is invoked after the coroutine is popped from the call stack.
      *
-     * @param StrandInterface $strand The strand that is executing the coroutine.
+     * @param Strand $strand The strand that is executing the coroutine.
      */
-    public function finalize(StrandInterface $strand)
+    public function finalize(Strand $strand)
     {
     }
 }