+++ src/Coroutine/StandardCoroutineAdaptor.php
@@ -6,7 +6,7 @@ use Generator;
 use Icecave\Repr\Repr;
 use InvalidArgumentException;
 use React\Promise\PromiseInterface;
-use Recoil\Kernel\Strand\StrandInterface;
+use Recoil\Kernel\Strand\Strand;
 use Recoil\Recoil;
 
 /**
@@ -17,13 +17,13 @@ class StandardCoroutineAdaptor implements CoroutineAdaptor
     /**
      * Adapt a value into a coroutine.
      *
-     * @param StrandInterface $strand The currently executing strand.
-     * @param mixed           $value  The value to adapt.
+     * @param Strand $strand The currently executing strand.
+     * @param mixed  $value  The value to adapt.
      *
      * @return Coroutine
      * @throws InvalidArgumentException if now valid adaptation can be made.
      */
-    public function adapt(StrandInterface $strand, $value)
+    public function adapt(Strand $strand, $value)
     {
         while ($value instanceof CoroutineProvider) {
             $value = $value->coroutine($strand);