+++ src/Kernel/Api/StandardKernelApi.php
@@ -3,7 +3,7 @@
 namespace Recoil\Kernel\Api;
 
 use Exception;
-use Recoil\Kernel\Strand\StrandInterface;
+use Recoil\Kernel\Strand\Strand;
 
 /**
  * The default kernel API implementation.
@@ -13,9 +13,9 @@ class StandardKernelApi implements KernelApi
     /**
      * Get the strand the coroutine is executing on.
      *
-     * @param StrandInterface $strand The currently executing strand.
+     * @param Strand $strand The currently executing strand.
      */
-    public function strand(StrandInterface $strand)
+    public function strand(Strand $strand)
     {
         $strand->resumeWithValue($strand);
     }
@@ -23,9 +23,9 @@ class StandardKernelApi implements KernelApi
     /**
      * Get the coroutine kernel that the current strand is executing on.
      *
-     * @param StrandInterface $strand The currently executing strand.
+     * @param Strand $strand The currently executing strand.
      */
-    public function kernel(StrandInterface $strand)
+    public function kernel(Strand $strand)
     {
         $strand->resumeWithValue($strand->kernel());
     }
@@ -33,9 +33,9 @@ class StandardKernelApi implements KernelApi
     /**
      * Get the React event-loop that the coroutine kernel is executing on.
      *
-     * @param StrandInterface $strand The currently executing strand.
+     * @param Strand $strand The currently executing strand.
      */
-    public function eventLoop(StrandInterface $strand)
+    public function eventLoop(Strand $strand)
     {
         $strand->resumeWithValue($strand->kernel()->eventLoop());
     }
@@ -43,10 +43,10 @@ class StandardKernelApi implements KernelApi
     /**
      * Return a value to the calling coroutine.
      *
-     * @param StrandInterface $strand The currently executing strand.
-     * @param mixed           $value  The value to send to the calling coroutine.
+     * @param Strand $strand The currently executing strand.
+     * @param mixed  $value  The value to send to the calling coroutine.
      */
-    public function return_(StrandInterface $strand, $value = null)
+    public function return_(Strand $strand, $value = null)
     {
         $strand->returnValue($value);
     }
@@ -54,10 +54,10 @@ class StandardKernelApi implements KernelApi
     /**
      * Throw an exception to the calling coroutine.
      *
-     * @param StrandInterface $strand    The currently executing strand.
-     * @param Exception       $exception The error to send to the calling coroutine.
+     * @param Strand    $strand    The currently executing strand.
+     * @param Exception $exception The error to send to the calling coroutine.
      */
-    public function throw_(StrandInterface $strand, Exception $exception)
+    public function throw_(Strand $strand, Exception $exception)
     {
         $strand->throwException($exception);
     }
@@ -69,10 +69,10 @@ class StandardKernelApi implements KernelApi
      * The callback is guaranteed to be called even if a reference to the
      * coroutine is held elsewhere (unlike a PHP finally block in a generator).
      *
-     * @param StrandInterface $strand   The currently executing strand.
-     * @param callable        $callback The callback to invoke.
+     * @param Strand   $strand   The currently executing strand.
+     * @param callable $callback The callback to invoke.
      */
-    public function finally_(StrandInterface $strand, callable $callback)
+    public function finally_(Strand $strand, callable $callback)
     {
         $strand->current()->registerFinalizeCallback($callback);
 
@@ -82,9 +82,9 @@ class StandardKernelApi implements KernelApi
     /**
      * Terminate execution of the strand.
      *
-     * @param StrandInterface $strand The currently executing strand.
+     * @param Strand $strand The currently executing strand.
      */
-    public function terminate(StrandInterface $strand)
+    public function terminate(Strand $strand)
     {
         $strand->terminate();
     }
@@ -92,10 +92,10 @@ class StandardKernelApi implements KernelApi
     /**
      * Suspend execution for a specified period of time.
      *
-     * @param StrandInterface $strand  The currently executing strand.
-     * @param number          $timeout The number of seconds to wait before resuming.
+     * @param Strand $strand  The currently executing strand.
+     * @param number $timeout The number of seconds to wait before resuming.
      */
-    public function sleep(StrandInterface $strand, $timeout)
+    public function sleep(Strand $strand, $timeout)
     {
         return new Sleep($timeout);
     }
@@ -103,9 +103,9 @@ class StandardKernelApi implements KernelApi
     /**
      * Suspend execution of the strand.
      *
-     * @param StrandInterface $strand The currently executing strand.
+     * @param Strand $strand The currently executing strand.
      */
-    public function suspend(StrandInterface $strand, callable $callback)
+    public function suspend(Strand $strand, callable $callback)
     {
         $strand->suspend();
 
@@ -118,11 +118,11 @@ class StandardKernelApi implements KernelApi
      * If the coroutine does not complete within the specified time it is
      * cancelled.
      *
-     * @param StrandInterface $strand    The currently executing strand.
-     * @param number          $timeout   The number of seconds to wait before cancelling.
-     * @param mixed           $coroutine The coroutine to execute.
+     * @param Strand $strand    The currently executing strand.
+     * @param number $timeout   The number of seconds to wait before cancelling.
+     * @param mixed  $coroutine The coroutine to execute.
      */
-    public function timeout(StrandInterface $strand, $timeout, $coroutine)
+    public function timeout(Strand $strand, $timeout, $coroutine)
     {
         return new Timeout($timeout, $coroutine);
     }
@@ -134,10 +134,10 @@ class StandardKernelApi implements KernelApi
      * are completed. If any of the coroutines fails, the remaining coroutines
      * are terminated.
      *
-     * @param StrandInterface $strand     The currently executing strand.
-     * @param array           $coroutines The coroutines to execute.
+     * @param Strand $strand     The currently executing strand.
+     * @param array  $coroutines The coroutines to execute.
      */
-    public function all(StrandInterface $strand, array $coroutines)
+    public function all(Strand $strand, array $coroutines)
     {
         return new WaitAll($coroutines);
     }
@@ -145,9 +145,9 @@ class StandardKernelApi implements KernelApi
     /**
      * Suspend the strand until the next tick.
      *
-     * @param StrandInterface $strand The currently executing strand.
+     * @param Strand $strand The currently executing strand.
      */
-    public function cooperate(StrandInterface $strand)
+    public function cooperate(Strand $strand)
     {
         $strand->suspend();
 
@@ -164,9 +164,9 @@ class StandardKernelApi implements KernelApi
     /**
      * Resume the strand immediately.
      *
-     * @param StrandInterface $strand The currently executing strand.
+     * @param Strand $strand The currently executing strand.
      */
-    public function noop(StrandInterface $strand)
+    public function noop(Strand $strand)
     {
         $strand->resumeWithValue(null);
     }
@@ -174,10 +174,10 @@ class StandardKernelApi implements KernelApi
     /**
      * Execute a coroutine on its own strand.
      *
-     * @param StrandInterface $strand    The currently executing strand.
-     * @param mixed           $coroutine The coroutine to execute.
+     * @param Strand $strand    The currently executing strand.
+     * @param mixed  $coroutine The coroutine to execute.
      */
-    public function execute(StrandInterface $strand, $coroutine)
+    public function execute(Strand $strand, $coroutine)
     {
         $substrand = $strand
             ->kernel()
@@ -189,10 +189,10 @@ class StandardKernelApi implements KernelApi
     /**
      * Wait for one or more of the given strands to exit.
      *
-     * @param StrandInterface   $strand  The currently executing strand.
-     * @param StrandInterface[] $strands The strands to wait for.
+     * @param Strand   $strand  The currently executing strand.
+     * @param Strand[] $strands The strands to wait for.
      */
-    public function select(StrandInterface $strand, array $strands)
+    public function select(Strand $strand, array $strands)
     {
         return new Select($strands);
     }
@@ -203,10 +203,10 @@ class StandardKernelApi implements KernelApi
      * The React event-loop can optionally be stopped when all strands have been
      * terminated.
      *
-     * @param StrandInterface $strand        The currently executing strand.
-     * @param boolean         $stopEventLoop Indicates whether or not the React event-loop should also be stopped.
+     * @param Strand  $strand        The currently executing strand.
+     * @param boolean $stopEventLoop Indicates whether or not the React event-loop should also be stopped.
      */
-    public function stop(StrandInterface $strand, $stopEventLoop = true)
+    public function stop(Strand $strand, $stopEventLoop = true)
     {
         $strand->terminate();