+++ src/Kernel/Strand/StackBase.php
@@ -21,9 +21,9 @@ class StackBase implements Coroutine
      *
      * @codeCoverageIgnore
      *
-     * @param StrandInterface $strand The strand that is executing the coroutine.
+     * @param Strand $strand The strand that is executing the coroutine.
      */
-    public function call(StrandInterface $strand)
+    public function call(Strand $strand)
     {
         throw new LogicException('Not supported.');
     }
@@ -31,10 +31,10 @@ class StackBase implements Coroutine
     /**
      * Resume execution of a suspended coroutine by passing it a value.
      *
-     * @param StrandInterface $strand The strand that is executing the coroutine.
-     * @param mixed           $value  The value to send to the coroutine.
+     * @param Strand $strand The strand that is executing the coroutine.
+     * @param mixed  $value  The value to send to the coroutine.
      */
-    public function resumeWithValue(StrandInterface $strand, $value)
+    public function resumeWithValue(Strand $strand, $value)
     {
         $strand->emit('success', [$strand, $value]);
         $strand->emit('exit', [$strand]);
@@ -47,10 +47,10 @@ class StackBase implements Coroutine
     /**
      * Resume execution of a suspended coroutine by passing it an exception.
      *
-     * @param StrandInterface $strand    The strand that is executing the coroutine.
-     * @param Exception       $exception The exception to send to the coroutine.
+     * @param Strand    $strand    The strand that is executing the coroutine.
+     * @param Exception $exception The exception to send to the coroutine.
      */
-    public function resumeWithException(StrandInterface $strand, Exception $exception)
+    public function resumeWithException(Strand $strand, Exception $exception)
     {
         $throwException = true;
 
@@ -73,9 +73,9 @@ class StackBase implements Coroutine
     /**
      * Inform the coroutine that the executing strand is being terminated.
      *
-     * @param StrandInterface $strand The strand that is executing the coroutine.
+     * @param Strand $strand The strand that is executing the coroutine.
      */
-    public function terminate(StrandInterface $strand)
+    public function terminate(Strand $strand)
     {
         $strand->emit('terminate', [$strand]);
         $strand->emit('exit', [$strand]);