+++ src/Kernel/Strand/StandardStrandFactory.php
@@ -14,7 +14,7 @@ class StandardStrandFactory implements StrandFactory
      *
      * @param Kernel The kernel on which the strand will execute.
      *
-     * @return StrandInterface
+     * @return Strand
      */
     public function createStrand(Kernel $kernel)
     {