+++ src/Recoil.php
@@ -5,7 +5,7 @@ namespace Recoil;
 use React\EventLoop\LoopInterface;
 use Recoil\Kernel\Api\KernelApiCall;
 use Recoil\Kernel\StandardKernel;
-use Recoil\Kernel\Strand\StrandInterface;
+use Recoil\Kernel\Strand\Strand;
 
 /**
  * Public facade for Kernel API calls.
@@ -31,7 +31,7 @@ use Recoil\Kernel\Strand\StrandInterface;
  * @method static noop() [COROUTINE] Resume the strand immediately.
  * @method static cooperate() [COROUTINE] Suspend the strand until the next tick.
  * @method static execute($coroutine) [COROUTINE] Execute a coroutine on its own strand.
- * @method static select(StrandInterface $strand, array $strands) [COROUTINE] Wait for one or more of the given strands to exit.
+ * @method static select(Strand $strand, array $strands) [COROUTINE] Wait for one or more of the given strands to exit.
  * @method static stop(bool $stopEventLoop = true) [COROUTINE] Stop the coroutine kernel / event-loop.
  */
 abstract class Recoil